[mod] stabilise terminal user input prompt
-requires a decent terminal implementing ansi code parsing
-note that the sequence arrived at is, to my mind, wrong. there seems
to be a disconnect between where python believes the cursor to be and
where it actually is. after the unavoidable (python built-in support
for terminals is rubbish) return / newline after single char read on
stdin, resetting should be: line up, clear to end, return, re-write
line. after a mass of trial and error, the additional '\033[B' (move
cursor down one row) escape 'fixed' the inability to stabilise the
position of the line. cannot comprehend why and it's certainly
inconsistant with my shell efforts